refactor: Split the File Provider relay and fetch-progress types out of the domain host
## Summary
- `FileProviderDomainHost.swift` had grown to 1,626 lines carrying three
  unrelated responsibilities: the domain host itself, the owner-side XPC relay,
  and the servicing progress push. Split the latter two into their own files so
  the upcoming Finder copy-dialog progress work lands in a reviewable diff
  rather than at the bottom of an already-unwieldy file.
- Pure move: no behavior change, no API change, no call-site change.

## Changes
- Add `KernovaKit/Sources/KernovaKit/FileProviderRelayService.swift` — the
  `// MARK: - Relay service` section (`FileProviderRelayService`) moved verbatim,
  with a file-header comment in the style of `FileProviderDomainHost.swift`'s.
- Add `KernovaKit/Sources/KernovaKit/FileProviderFetchProgress.swift` — the
  `// MARK: - Servicing progress push` section (`FetchProgressThrottle`,
  `FetchProgressPusher`) moved verbatim, likewise headed.
- Trim `FileProviderDomainHost.swift` to the domain host itself.
- No `project.pbxproj` edit needed: KernovaKit is a SwiftPM package, so new
  sources are picked up automatically.

## Test plan
- [x] `make build` succeeds
- [x] `make test` passes (all three test targets)
- [x] `make lint` clean
- [x] `make test-package` passes (339 tests, 39 suites)
